+++ changelog.md
@@ -97,6 +97,7 @@ Codion Change Log
 - LabelBuilder, overloaded builder() methods removed.
 - BorderLayoutPanelBuilder, overloaded builder() methods removed.
 - PanelBuilder, overloaded builder() methods removed.
+- ProgressBarBuilder, overloaded builder() methods removed.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui